Updated Plugins Page. Moved Settings and related Pages to Settings Folder.
