+++ LlamaRun/LlamaRun.vcxproj
@@ -131,8 +131,8 @@
       <DependentUpon>FirstRunWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </ClInclude>
-    <ClInclude Include="HomePage_SettingsWindow.xaml.h">
-      <DependentUpon>HomePage_SettingsWindow.xaml</DependentUpon>
+    <ClInclude Include="Settings/HomePage_SettingsWindow.xaml.h">
+      <DependentUpon>Settings/HomePage_SettingsWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </ClInclude>
     <ClInclude Include="pch.h" />
@@ -144,11 +144,11 @@
     </ClInclude>
     <ClInclude Include="PluginDef.h" />
     <ClInclude Include="PluginManager.h" />
-    <ClInclude Include="PluginPage_SettingsWindow.xaml.h">
-      <DependentUpon>PluginPage_SettingsWindow.xaml</DependentUpon>
+    <ClInclude Include="Settings/PluginPage_SettingsWindow.xaml.h">
+      <DependentUpon>Settings/PluginPage_SettingsWindow.xaml</DependentUpon>
     </ClInclude>
-    <ClInclude Include="SettingsWindow.xaml.h">
-      <DependentUpon>SettingsWindow.xaml</DependentUpon>
+    <ClInclude Include="Settings/SettingsWindow.xaml.h">
+      <DependentUpon>Settings/SettingsWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </ClInclude>
     <ClInclude Include="ShowOllamaUnavaliableWindow.h" />
@@ -159,12 +159,12 @@
     <Page Include="FirstRunWindow.xaml">
       <SubType>Designer</SubType>
     </Page>
-    <Page Include="HomePage_SettingsWindow.xaml">
+    <Page Include="Settings/HomePage_SettingsWindow.xaml">
       <SubType>Designer</SubType>
     </Page>
     <Page Include="MainWindow.xaml" />
-    <Page Include="PluginPage_SettingsWindow.xaml" />
-    <Page Include="SettingsWindow.xaml">
+    <Page Include="Settings/PluginPage_SettingsWindow.xaml" />
+    <Page Include="Settings/SettingsWindow.xaml">
       <SubType>Designer</SubType>
     </Page>
   </ItemGroup>
@@ -174,8 +174,8 @@
       <DependentUpon>FirstRunWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </ClCompile>
-    <ClCompile Include="HomePage_SettingsWindow.xaml.cpp">
-      <DependentUpon>HomePage_SettingsWindow.xaml</DependentUpon>
+    <ClCompile Include="Settings/HomePage_SettingsWindow.xaml.cpp">
+      <DependentUpon>Settings/HomePage_SettingsWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </ClCompile>
     <ClCompile Include="pch.cpp">
@@ -190,11 +190,11 @@
     <ClCompile Include="$(GeneratedFilesDir)module.g.cpp" />
     <ClCompile Include="PluginDef.cpp" />
     <ClCompile Include="PluginManager.cpp" />
-    <ClCompile Include="PluginPage_SettingsWindow.xaml.cpp">
-      <DependentUpon>PluginPage_SettingsWindow.xaml</DependentUpon>
+    <ClCompile Include="Settings/PluginPage_SettingsWindow.xaml.cpp">
+      <DependentUpon>Settings/PluginPage_SettingsWindow.xaml</DependentUpon>
     </ClCompile>
-    <ClCompile Include="SettingsWindow.xaml.cpp">
-      <DependentUpon>SettingsWindow.xaml</DependentUpon>
+    <ClCompile Include="Settings/SettingsWindow.xaml.cpp">
+      <DependentUpon>Settings/SettingsWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </ClCompile>
     <ClCompile Include="ShowOllamaUnavaliableWindow.cpp" />
@@ -205,20 +205,20 @@
       <DependentUpon>FirstRunWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Midl>
-    <Midl Include="HomePage_SettingsWindow.idl">
-      <DependentUpon>HomePage_SettingsWindow.xaml</DependentUpon>
+    <Midl Include="Settings/HomePage_SettingsWindow.idl">
+      <DependentUpon>Settings/HomePage_SettingsWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Midl>
     <Midl Include="MainWindow.idl">
       <SubType>Code</SubType>
       <DependentUpon>MainWindow.xaml</DependentUpon>
     </Midl>
-    <Midl Include="PluginPage_SettingsWindow.idl">
-      <DependentUpon>PluginPage_SettingsWindow.xaml</DependentUpon>
+    <Midl Include="Settings/PluginPage_SettingsWindow.idl">
+      <DependentUpon>Settings/PluginPage_SettingsWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Midl>
-    <Midl Include="SettingsWindow.idl">
-      <DependentUpon>SettingsWindow.xaml</DependentUpon>
+    <Midl Include="Settings/SettingsWindow.idl">
+      <DependentUpon>Settings/SettingsWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Midl>
   </ItemGroup>
