--- LlamaRun/PluginPage_SettingsWindow.xaml.cpp
@@ -1,29 +1,0 @@
-#include "pch.h"
-#include "PluginPage_SettingsWindow.xaml.h"
-#if __has_include("PluginPage_SettingsWindow.g.cpp")
-#include "PluginPage_SettingsWindow.g.cpp"
-#endif
-
-using namespace winrt;
-using namespace Microsoft::UI::Xaml;
-
-// To learn more about WinUI, the WinUI project structure,
-// and more about our project templates, see: http://aka.ms/winui-project-info.
-
-namespace winrt::LlamaRun::implementation
-{
-    int32_t PluginPage_SettingsWindow::MyProperty()
-    {
-        throw hresult_not_implemented();
-    }
-
-    void PluginPage_SettingsWindow::MyProperty(int32_t /* value */)
-    {
-        throw hresult_not_implemented();
-    }
-
-    void PluginPage_SettingsWindow::myButton_Click(IInspectable const&, RoutedEventArgs const&)
-    {
-        myButton().Content(box_value(L"Clicked"));
-    }
-}
