+++ LlamaRun/Settings/SettingsWindow.xaml
@@ -13,26 +13,17 @@
     </Window.SystemBackdrop>
 
     <Grid x:Name="rootPanel">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="auto"/>
-        </Grid.RowDefinitions>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="*" />
-        </Grid.ColumnDefinitions>
-        <NavigationView Grid.Column="0" x:Name="NavView" Loaded="NavigationView_Loaded" SelectionChanged="NavigationView_SelectionChanged">
+        <NavigationView x:Name="NavView" IsBackButtonVisible="Collapsed" IsSettingsVisible="False" Loaded="NavigationView_Loaded" SelectionChanged="NavigationView_SelectionChanged">
             <NavigationView.MenuItems>
                 <NavigationViewItem Content="Home" Tag="Home" Icon="Home"/>
                 <NavigationViewItem Content="Plugins" Tag="Plugins" Icon="Repair"/>
             </NavigationView.MenuItems>
 
-        </NavigationView>
-        <StackPanel Grid.Row="0" Orientation="Vertical" HorizontalAlignment="Center" VerticalAlignment="Center">
             <Frame x:Name="ContentFrame"/>
-        </StackPanel>
-        <StackPanel Grid.Row="1" Margin="0,-100,0,0" Orientation="Horizontal" HorizontalAlignment="Center" VerticalAlignment="Center">
+        </NavigationView>
+        <StackPanel Margin="0, 0, 0, 20" Orientation="Horizontal" HorizontalAlignment="Center" VerticalAlignment="Bottom">
             <TextBlock Margin="0,0,0,0" FontSize="30">Made With</TextBlock>
-            <Image Source="Assets/red_heart_animated.gif" Margin="0,0,0,-10" Width="50"></Image>
+            <Image Source="../Assets/red_heart_animated.gif" Margin="0,0,0,-10" Width="50"></Image>
             <TextBlock FontSize="30">By&#160;</TextBlock>
             <HyperlinkButton Margin="0,0,0,0" NavigateUri="https://github.com/krishbaidya" FontSize="30" Style="{StaticResource TextBlockStyleHyperlink}">Krish Baidya</HyperlinkButton>
         </StackPanel>
